[Source] Split Context into multiple files.
