+++ include/cocaine/context/signal.hpp
@@ -18,8 +18,8 @@
     along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-#ifndef COCAINE_SIGNAL_HPP
-#define COCAINE_SIGNAL_HPP
+#ifndef COCAINE_CONTEXT_SIGNAL_HPP
+#define COCAINE_CONTEXT_SIGNAL_HPP
 
 #include "cocaine/tuple.hpp"
